fix: handle duplicate error instead of sql error fix: set fqdns to null if you delete an app or a serviceapp fix: make stucked resources a separate command
