+++ app/Models/Service.php
@@ -2,13 +2,11 @@
 
 namespace App\Models;
 
-use App\Actions\Service\DeleteService;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
-use Symfony\Component\Yaml\Yaml;
 
 class Service extends BaseModel
 {
