+++ app/Models/ServiceApplication.php
@@ -5,7 +5,6 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use Illuminate\Support\Facades\Cache;
 
 class ServiceApplication extends BaseModel
 {
@@ -15,6 +14,7 @@ class ServiceApplication extends BaseModel
     protected static function booted()
     {
         static::deleting(function ($service) {
+            $service->update(['fqdn' => null]);
             $service->persistentStorages()->delete();
             $service->fileStorages()->delete();
         });
